+++ src/models/deviceSettings/tvSettings.ts
@@ -1,0 +1,6 @@
+import { DeviceSettings } from './deviceSettings';
+import { Daytime } from '../daytime';
+
+export class TvSettings extends DeviceSettings {
+  public automaticTurnOff: Daytime | undefined;
+}