+++ src/server/devices/DeviceCapability.ts
@@ -16,6 +16,7 @@ export enum DeviceCapability {
   speaker = 14,
   handleSensor = 15,
   batteryDriven = 16,
+  tv = 17,
   bluetoothDetector = 101,
   trackableDevice = 102,
 }