+++ src/server/devices/baseDeviceInterfaces/iTvDevice.ts
@@ -1,0 +1,13 @@
+import { iRoomDevice } from './iRoomDevice';
+
+export interface iTvDevice extends iRoomDevice {
+  readonly on: boolean;
+
+  turnOn(): void;
+
+  turnOff(): void;
+
+  volumeUp(): void;
+
+  volumeDown(): void;
+}