+++ src/server/devices/tv/index.ts
@@ -1,0 +1,2 @@
+export * from './tvDevice';
+export * from './tvDeviceType';