+++ src/server/devices/tv/tvDeviceType.ts
@@ -1,0 +1,4 @@
+export enum TvDeviceType {
+  Unknown,
+  Samsung,
+}